# spec/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Shift Aggregation API
  version: 1.0.0
  description: |
    月次シフト希望集約システムのAPI仕様。
    Google OAuthログイン後のセッション/JWTを前提とする。
servers:
  - url: /api/v1
...
paths:
...
  /auth/google/callback:
    get:
      tags:
        - Auth
      summary: OAuthコールバック
      security: []
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
      responses:
        '302':
          description: 認証成功後に画面へリダイレクト
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: ログアウト
      security:
        - sessionAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags:
        - Auth
      summary: ログインユーザー情報取得
      security:
        - sessionAuth: []
      responses:
        '200':
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Ok:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
...
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    SuccessResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
    User:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - employee
            - admin
        employeeId:
          type:
            - integer
            - 'null'
          format: int64
...
    MeResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/User'
...